# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectApplicationType = typeof ProjectApplicationType[keyof typeof ProjectApplicationType];


export const ProjectApplicationType = {
  ecommerce: 'ecommerce',
  news: 'news',
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  paused: 'paused',
} as const;

export interface Project {
  id: string;
  name: string;
  description: string;
  applicationType: ProjectApplicationType;
  applicationUrl: string;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputApplicationType = typeof ProjectInputApplicationType[keyof typeof ProjectInputApplicationType];


export const ProjectInputApplicationType = {
  ecommerce: 'ecommerce',
  news: 'news',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description: string;
  applicationType: ProjectInputApplicationType;
  applicationUrl: string;
}

export type ProjectUpdateApplicationType = typeof ProjectUpdateApplicationType[keyof typeof ProjectUpdateApplicationType];


export const ProjectUpdateApplicationType = {
  ecommerce: 'ecommerce',
  news: 'news',
} as const;

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  paused: 'paused',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  applicationType?: ProjectUpdateApplicationType;
  applicationUrl?: string;
  status?: ProjectUpdateStatus;
}

export interface DashboardSummary {
  projectCount: number;
  testRunCount: number;
  /** @nullable */
  passRate: number | null;
  violationCount: number;
}

export type AdaptiveContractProfile = typeof AdaptiveContractProfile[keyof typeof AdaptiveContractProfile];


export const AdaptiveContractProfile = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AdaptiveContract {
  id: string;
  projectId: string;
  profile: AdaptiveContractProfile;
  networkProfile: string;
  imagePolicy: string;
  javascriptPolicy: string;
  featurePolicy: string;
  /** @minimum 1 */
  maxResourceSizeKb: number;
  /** @minimum 1 */
  maxLcpMs: number;
  createdAt: string;
  updatedAt: string;
}

export type AdaptiveContractInputProfile = typeof AdaptiveContractInputProfile[keyof typeof AdaptiveContractInputProfile];


export const AdaptiveContractInputProfile = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AdaptiveContractInput {
  profile: AdaptiveContractInputProfile;
  /** @minLength 1 */
  networkProfile: string;
  /** @minLength 1 */
  imagePolicy: string;
  /** @minLength 1 */
  javascriptPolicy: string;
  /** @minLength 1 */
  featurePolicy: string;
  /** @minimum 1 */
  maxResourceSizeKb: number;
  /** @minimum 1 */
  maxLcpMs: number;
}

export type TestProfileKey = typeof TestProfileKey[keyof typeof TestProfileKey];


export const TestProfileKey = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TestProfile {
  key: TestProfileKey;
  name: string;
  description: string;
  networkProfile: string;
  imagePolicy: string;
  javascriptPolicy: string;
  featurePolicy: string;
  maxResourceSizeKb: number;
  maxLcpMs: number;
}

export type TestRunProfile = typeof TestRunProfile[keyof typeof TestRunProfile];


export const TestRunProfile = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TestRunMethod = typeof TestRunMethod[keyof typeof TestRunMethod];


export const TestRunMethod = {
  adaptive_behavior: 'adaptive_behavior',
  performance: 'performance',
  resource: 'resource',
  full_validation: 'full_validation',
} as const;

export type TestRunConfiguration = { [key: string]: unknown };

export type TestRunStatus = typeof TestRunStatus[keyof typeof TestRunStatus];


export const TestRunStatus = {
  queued: 'queued',
  running: 'running',
  passed: 'passed',
  violated: 'violated',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface TestRun {
  id: string;
  projectId: string;
  profile: TestRunProfile;
  method: TestRunMethod;
  configuration: TestRunConfiguration;
  status: TestRunStatus;
  createdAt: string;
  updatedAt: string;
}

export type TestRunInputProfile = typeof TestRunInputProfile[keyof typeof TestRunInputProfile];


export const TestRunInputProfile = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TestRunInputMethod = typeof TestRunInputMethod[keyof typeof TestRunInputMethod];


export const TestRunInputMethod = {
  adaptive_behavior: 'adaptive_behavior',
  performance: 'performance',
  resource: 'resource',
  full_validation: 'full_validation',
} as const;

export type TestRunInputConfiguration = { [key: string]: unknown };

export interface TestRunInput {
  profile: TestRunInputProfile;
  method: TestRunInputMethod;
  configuration?: TestRunInputConfiguration;
}