# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: External systems under test
  - name: dashboard
    description: Dashboard summary operations
  - name: contracts
    description: Adaptive contract configuration
  - name: tests
    description: Test profiles and queued test runs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      description: Returns the current project and testing summary for the workspace
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid project
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Project deleted
        "404":
          description: Project not found
  /projects/{id}/contract:
    get:
      operationId: getAdaptiveContract
      tags: [contracts]
      summary: Get a project's adaptive contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Adaptive contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveContract"
        "404":
          description: Project or contract not found
    put:
      operationId: upsertAdaptiveContract
      tags: [contracts]
      summary: Create or update a project's adaptive contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdaptiveContractInput"
      responses:
        "200":
          description: Saved adaptive contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveContract"
        "400":
          description: Invalid adaptive contract
        "404":
          description: Project not found
  /test-profiles:
    get:
      operationId: listTestProfiles
      tags: [tests]
      summary: List reusable test profiles
      responses:
        "200":
          description: Reusable test profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestProfile"
  /projects/{id}/tests:
    get:
      operationId: listTestRuns
      tags: [tests]
      summary: List a project's test runs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestRun"
        "404":
          description: Project not found
    post:
      operationId: createTestRun
      tags: [tests]
      summary: Queue a test run
      description: Creates a queued test run without executing a browser test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestRunInput"
      responses:
        "201":
          description: Queued test run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestRun"
        "400":
          description: Invalid test configuration
        "404":
          description: Project not found
  /projects/{id}/tests/{testId}:
    get:
      operationId: getTestRun
      tags: [tests]
      summary: Get a test run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: testId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestRun"
        "404":
          description: Test run not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        applicationType:
          type: string
          enum: [ecommerce, news]
        applicationUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, paused]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - applicationType
        - applicationUrl
        - status
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        applicationType:
          type: string
          enum: [ecommerce, news]
        applicationUrl:
          type: string
          format: uri
      required:
        - name
        - description
        - applicationType
        - applicationUrl
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        applicationType:
          type: string
          enum: [ecommerce, news]
        applicationUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, paused]
    DashboardSummary:
      type: object
      properties:
        projectCount:
          type: integer
        testRunCount:
          type: integer
        passRate:
          type: ["number", "null"]
        violationCount:
          type: integer
      required:
        - projectCount
        - testRunCount
        - passRate
        - violationCount
    AdaptiveContract:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        profile:
          type: string
          enum: [low, medium, high]
        networkProfile:
          type: string
        imagePolicy:
          type: string
        javascriptPolicy:
          type: string
        featurePolicy:
          type: string
        maxResourceSizeKb:
          type: integer
          minimum: 1
        maxLcpMs:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - profile
        - networkProfile
        - imagePolicy
        - javascriptPolicy
        - featurePolicy
        - maxResourceSizeKb
        - maxLcpMs
        - createdAt
        - updatedAt
    AdaptiveContractInput:
      type: object
      properties:
        profile:
          type: string
          enum: [low, medium, high]
        networkProfile:
          type: string
          minLength: 1
        imagePolicy:
          type: string
          minLength: 1
        javascriptPolicy:
          type: string
          minLength: 1
        featurePolicy:
          type: string
          minLength: 1
        maxResourceSizeKb:
          type: integer
          minimum: 1
        maxLcpMs:
          type: integer
          minimum: 1
      required:
        - profile
        - networkProfile
        - imagePolicy
        - javascriptPolicy
        - featurePolicy
        - maxResourceSizeKb
        - maxLcpMs
    TestProfile:
      type: object
      properties:
        key:
          type: string
          enum: [low, medium, high]
        name:
          type: string
        description:
          type: string
        networkProfile:
          type: string
        imagePolicy:
          type: string
        javascriptPolicy:
          type: string
        featurePolicy:
          type: string
        maxResourceSizeKb:
          type: integer
        maxLcpMs:
          type: integer
      required:
        - key
        - name
        - description
        - networkProfile
        - imagePolicy
        - javascriptPolicy
        - featurePolicy
        - maxResourceSizeKb
        - maxLcpMs
    TestRun:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        profile:
          type: string
          enum: [low, medium, high]
        method:
          type: string
          enum: [adaptive_behavior, performance, resource, full_validation]
        configuration:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [queued, running, passed, violated, failed, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - profile
        - method
        - configuration
        - status
        - createdAt
        - updatedAt
    TestRunInput:
      type: object
      properties:
        profile:
          type: string
          enum: [low, medium, high]
        method:
          type: string
          enum: [adaptive_behavior, performance, resource, full_validation]
        configuration:
          type: object
          additionalProperties: true
      required:
        - profile
        - method